@@ -12,11 +12,8 @@ use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @ORM\Table(name="`user`")
- *
- * @ORM\Entity()
- */
+#[ORM\Entity()]
+#[ORM\Table(name: "user")]
 class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUserInterface
 {
     /**
@@ -25,11 +22,9 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups("public")
      *
      * @OA\Property(type="guid")
-     *
-     * @ORM\Column(type="guid")
-     *
-     * @ORM\Id
      */
+    #[ORM\Id]
+    #[ORM\Column(type: "guid")]
     private $id;
 
     /**
@@ -38,9 +33,8 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Assert\Type("integer")
      *
      * @OA\Property(type="integer")
-     *
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Column(type: "integer")]
     private $createdAt;
 
     /**
@@ -49,9 +43,8 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups({"read_me", "write_user"})
      *
      * @OA\Property(type="string")
-     *
-     * @ORM\Column(type="string", unique=true)
      */
+    #[ORM\Column(type: "string", unique: true)]
     private $login;
 
     /**
@@ -60,61 +53,52 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups({"write_user"})
      *
      * @OA\Property(type="string")
-     *
-     * @ORM\Column(type="string")
      */
+    #[ORM\Column(type: "string")]
     private $password;
 
     /**
      * @Assert\NotBlank()
      *
      * @OA\Property(type="guid")
-     *
-     * @ORM\Column(type="guid", unique=true)
      */
+    #[ORM\Column(type: "guid", unique: true)]
     private $secretKey;
 
     /**
      * @Groups({"read_me"})
      *
      * @OA\Property(type="array", @OA\Items(type="App\Entity\Group"))
-     *
-     * @ORM\JoinTable(name="users_groups")
-     *
-     * @ORM\ManyToMany(targetEntity="App\Entity\Group", inversedBy="users")
      */
+    #[ORM\ManyToMany(targetEntity: "App\Entity\Group", inversedBy: "users")]
+    #[ORM\JoinTable(name: "users_groups")]
     private $groups;
 
     /**
      * @OA\Property(type="array", @OA\Items(type="App\Entity\Message"))
-     *
-     * @ORM\OneToMany(targetEntity="App\Entity\Message", mappedBy="author")
      */
+    #[ORM\OneToMany(targetEntity: "App\Entity\Message", mappedBy: "author")]
     private $messages;
 
     /**
      * @OA\Property(type="array", @OA\Items(type="App\Entity\Reaction"))
-     *
-     * @ORM\OneToMany(targetEntity="App\Entity\Reaction", mappedBy="author")
      */
+    #[ORM\OneToMany(targetEntity: "App\Entity\Reaction", mappedBy: "author")]
     private Collection $reactions;
 
     /**
      * @OA\Property(type="array", @OA\Items(type="App\Entity\Bookmark"))
-     *
-     * @ORM\OneToMany(targetEntity="App\Entity\Bookmark", mappedBy="user")
      */
+    #[ORM\OneToMany(targetEntity: "App\Entity\Bookmark", mappedBy: "user")]
     private $bookmarks;
 
     /**
      * @Groups({"read_me", "read_user", "write_user", "read_message_preview"})
      *
      * @OA\Property(type="App\Entity\File")
-     *
-     * @ORM\JoinColumn(name="avatar_id", referencedColumnName="id")
-     *
-     * @ORM\OneToOne(targetEntity="App\Entity\File")
      */
+    #[ORM\OneToOne(targetEntity: "App\Entity\File")]
+    #[ORM\JoinColumn(name: "avatar_id", referencedColumnName: "id")]
     private $avatar;
 
     /**
@@ -123,9 +107,8 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups("public")
      *
      * @OA\Property(type="string")
-     *
-     * @ORM\Column(type="string")
      */
+    #[ORM\Column(type: "string")]
     private $name;
 
     /**
@@ -134,18 +117,15 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups({"read_me", "read_user", "write_user"})
      *
      * @OA\Property(type="object")
-     *
-     * @ORM\Column(type="json", nullable=true)
      */
+    #[ORM\Column(type: "json", nullable: true)]
     private $data;
 
     /**
      * @OA\Property(type="array", @OA\Items(type="App\Entity\Notification"))
-     *
-     * @ORM\OneToMany(targetEntity="App\Entity\Notification", mappedBy="owner")
-     *
-     * @ORM\OrderBy({"createdAt" = "DESC"})
      */
+    #[ORM\OneToMany(targetEntity: "App\Entity\Notification", mappedBy: "owner")]
+    #[ORM\OrderBy(["createdAt" => "DESC"])]
     private $notifications;
 
     /**
@@ -154,9 +134,8 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups({"read_me"})
      *
      * @OA\Property(type="integer")
-     *
-     * @ORM\Column(type="integer", nullable=true)
      */
+    #[ORM\Column(type: "integer", nullable: true)]
     private $lastActivityDate;
 
     /**
@@ -165,9 +144,8 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
      * @Groups({"read_me"})
      *
      * @OA\Property(type="integer")
-     *
-     * @ORM\Column(type="integer", nullable=true)
      */
+    #[ORM\Column(type: "integer", nullable: true)]
     private int $lastNotificationEmailCheck;
 
     /**
@@ -370,7 +348,7 @@ class User extends ApiEntity implements UserInterface, PasswordAuthenticatedUser
     }
 
     // necessary for UserInterface
-    public function eraseCredentials()
+    public function eraseCredentials(): void
     {
     }
 }
